tko.bind) Fix circular dependency for `BindingHandler` / `applyBindings`
